requirements.txt
Flask
spotdl
python-dotenv
psycopg2-binary